requirements.txt
python-docx==1.1.2
pandas==2.2.3
openai>=2.0.0
